+++ src/main/java/ic3/analyticsops/common/AOPause.java
@@ -1,0 +1,10 @@
+package ic3.analyticsops.common;
+
+public abstract class AOPause
+{
+    public abstract long pauseMS();
+
+    @Override
+    public abstract String toString();
+
+}